feat: Support large DATA, progress, fetch slot & upload
Change FastbootResponse.DataSize from int to long and parse DATA length with long.Parse to handle large (>=2GB) transfers. Update read/upload logic to use long-sized counters while casting when calling USB read. Add CLI progress output via DataTransferProgressChanged to show percentage and byte counts. Make Fetch respect slot-aware partitions by appending current slot when applicable and add an Upload(filename, outputPath) helper to run legacy upload commands. Update tests (rename and expected size) to validate large DATA handling.

+++ SharpFastboot/DataModel/FastbootResponse.cs
@@ -5,7 +5,7 @@
         public FastbootState Result { get; set; }
         public string Response { get; set; } = "";
         public byte[]? Data { get; set; }
-        public int DataSize { get; set; }
+        public long DataSize { get; set; }
         public List<string> Info { get; set; } = new List<string>();
         public string Text { get; set; } = "";
         public string? Hash { get; set; }
